pyproject.toml
[tool.poetry]
name = "personal-bible-study-app"
version = "0.1.0"
description = ""
authors = ["sungho baek <ashsh0412@gmail.com>"]
readme = "README.md"
packages = [{include = "personal_bible_study_app"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "3.2.13"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
